requirements.txt
Pillow==11.0.0
pillow_heif==0.21.0
Requests==2.32.3
